backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
sentence-transformers==3.3.1
faiss-cpu==1.9.0
numpy==1.26.4
python-multipart==0.0.20
